Initial jeep-air-down
This is a access-point mode ESP32 project for managing the tire pressure of an
off-road vehicle. I'll include a wiring diagram later.

A pressure sensor measures the pressure of the tires through the fill hoses.
One solenoid connects the tires to the compressed air source. The other vents
the air from the tires. With this the user can connect the tires to the system
and focus on something else while the tire pressures are adjusted.

Basic dialog at 192.168.4.1

Signed-off-by: Joe Hershberger <joe.hershberger@gmail.com>