Make air-up and air-down into a restful API
Track the status in an async task. Make it possible to query the status
from the device.
Make the javascript poll the command status.
Don't allow starting a new command while the opposite is running.
Use the button style to indicate the command is active.
Get rid of the dialog box.
Allow for command cancellation.

Signed-off-by: Joe Hershberger <joe.hershberger@gmail.com>