Add a learning algorithm for approaching the desired pressure quickly
We can only measure the pressure when it's not changing, so the more time we
spend checking, the less time we spend adjusting. To optimize, we need to
guess as accurately as possible how long the valve should be open for a given
operation and ideally get to the desired pressure in one try. In reality, we
need measure some number fo times. We don't want to overshoot, but we want to
check as few times as possible.

The learning algorithm will observe how fast we are changing pressure and
extrapolate how much more time will be needed to guess somewhat accurately.

Signed-off-by: Joe Hershberger <joe.hershberger@gmail.com>