Improve UI handling in the javascript
Separate the updating of the UI from the place the changes come from. Also add
a task to poll the firmware for state changes. Now that the actions can be
initiated from buttons on the device, we should check the status all the time
and not just when we started an action in the web.

Also eliminate flicker in the buttons by tracking clicks in the UI and holding
that representation until the status catches up. By doing that we can clearly
convey that the command was received, and also when the device took action on
that command.

Signed-off-by: Joe Hershberger <joe.hershberger@gmail.com>